refactor(core): move address logic to crate
