+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "crates/astria-conductor",
   "crates/astria-config",
   "crates/astria-core",
+  "crates/astria-core-address",
   "crates/astria-core-consts",
   "crates/astria-core-crypto",
   "crates/astria-eyre",
@@ -36,6 +37,7 @@ default-members = [
   "crates/astria-conductor",
   "crates/astria-config",
   "crates/astria-core",
+  "crates/astria-core-address",
   "crates/astria-core-consts",
   "crates/astria-core-crypto",
   "crates/astria-grpc-mock",
