+++ crates/astria-core-address/CHANGELOG.md
@@ -1,0 +1,14 @@
+<!-- markdownlint-disable no-duplicate-heading -->
+
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased
+
+### Added
+
+- Initial release. [#1802](https://github.com/astriaorg/astria/pull/1802)
